improved menu
fix the comments. move the menu to the left. reduce the logo.

Remove the active class in the span tag

correct the names of the classes.

Rename the class to html. Align the lines in header.scss.

wrap the picture in a div.

Change the name of the classes, moved the block to the navigation block.

move the exit menu to the left, move the navigation blocks separately from the saidbar.

rename the file name, make the navigation block, change the structure of blocks in navigation.

split the file into saidebar and navigation